+++ derive/src/context/std/str.rs
@@ -10,7 +10,7 @@ impl Context {
 
         let Context {
             modulus,
-            implementation,
+            cartridge,
             std,
             modtype,
             struct_ident,
@@ -30,7 +30,7 @@ impl Context {
 
                 #[inline]
                 fn from_str(s: &str) -> #std::result::Result<Self, #std::num::ParseIntError> {
-                    let #field_ident = <#implementation as #modtype::Impl>::from_str(s, #modulus)?;
+                    let #field_ident = <#cartridge as #modtype::Cartridge>::from_str(s, #modulus)?;
                     #std::result::Result::Ok(#struct_expr)
                 }
             }